Második átalakítás - saját modulba helyezzük a függvényt (de van benne egy hiba)
